Fix a crash when an  Executor wrapped fork exit.
Fix #51298

forking inside an Execution Wrapper crashes when running the completion callbacks.

Rails 7.0 was not Execution wrapping the `runner` command.
Rails 7.2 changed the definition of `active_connection?`, the new definition doesn't contain the bug.

Therefore forking inside a script intended to be run with the `runner` command on 7.1 crashes. (see #51298)