Switch to Active Storage Azure service backend to azure-blob
Fix #49983

`azure-storage` (including `azure-storage-blob`) is deprecated and
will be retired permanently on September 13, 2024.

`azure-storage-blob` is quite large and depends on outdated gems.

This change replaces `azure-storage-blob` with `azure-blob`, a
smaller gem specifically designed for Active Storage.

It depends only on default and bundled gems and includes only the
features needed by Active Storage, making it easier to maintain.